NFT Integration Tests (#143)
* separate testing modules into separate files

* mint to alice - alice rejects

* mint to token contract itself

* alice mints to herself - burns

* mint to alice, hook transfers to bob

* alice mints to bob who rejects

* token contract transfer tests

* two actor tests completed

* fix dependencies

* don't link nft_actor into integration tests
