+++ testing/fil_token_integration/actors/basic_nft_actor/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2021"
 
 [dependencies]
 cid = { version = "0.8.3", default-features = false, features = ["serde-codec"] }
-frc46_token = { path = "../../../../frc46_token" }
 frcxx_nft = { path = "../../../../frcxx_nft" }
 frc42_dispatch = { path = "../../../../frc42_dispatch" }
 fvm_actor_utils = { version = "0.1.0", path = "../../../../fvm_actor_utils" }
